ref chameleon-system/chameleon-system#2: revert TCMSMediaFieldMapper changes, decorate mapper instead in MediaManagerBundle to switch the button target
