+++ src/CoreBundle/private/library/classes/mappers/TCMSMediaFieldMapper.class.php
@@ -9,33 +9,8 @@
  * file that was distributed with this source code.
  */
 
-use ChameleonSystem\CoreBundle\Interfaces\MediaManagerUrlGeneratorInterface;
-use Symfony\Component\Translation\TranslatorInterface;
-
 class TCMSMediaFieldMapper extends AbstractViewMapper
 {
-    /**
-     * @var MediaManagerUrlGeneratorInterface
-     */
-    private $mediaManagerUrlGenerator;
-
-    /**
-     * @var TranslatorInterface
-     */
-    private $translator;
-
-    /**
-     * @param MediaManagerUrlGeneratorInterface $mediaManagerUrlGenerator
-     * @param TranslatorInterface               $translator
-     */
-    public function __construct(
-        MediaManagerUrlGeneratorInterface $mediaManagerUrlGenerator,
-        TranslatorInterface $translator)
-    {
-        $this->mediaManagerUrlGenerator = $mediaManagerUrlGenerator;
-        $this->translator = $translator;
-    }
-
     /**
      * {@inheritdoc}
      */
@@ -52,33 +27,16 @@ class TCMSMediaFieldMapper extends AbstractViewMapper
      */
     public function Accept(IMapperVisitorRestricted $oVisitor, $bCachingEnabled, IMapperCacheTriggerRestricted $oCacheTriggerManager)
     {
+        /** @var $oTextModuleConfiguration TdbCmsTblConf */
         $sFieldName = $oVisitor->GetSourceObject('sFieldName');
         $sHtmlHiddenFields = $oVisitor->GetSourceObject('sHtmlHiddenFields');
         $sRecordId = $oVisitor->GetSourceObject('sRecordId');
         $sTableId = $oVisitor->GetSourceObject('sTableId');
-
-        $mediaManagerButtonHtml = $this->getMediaManagerButtonHtml();
-        $oVisitor->SetMappedValue('sHtmlManageMediaButton', $mediaManagerButtonHtml);
+        $sHTMLManageMediaButton = TCMSRender::DrawButton(TGlobal::Translate('chameleon_system_core.link.open_media_manager'), "javascript:loadMediaManager('".$sRecordId."','".$sTableId."','".$sFieldName."');", URL_CMS.'/images/icons/image.gif');
+        $oVisitor->SetMappedValue('sHtmlManageMediaButton', $sHTMLManageMediaButton);
         $oVisitor->SetMappedValue('sFieldName', $sFieldName);
         $oVisitor->SetMappedValue('sRecordId', $sRecordId);
         $oVisitor->SetMappedValue('sTableId', $sTableId);
         $oVisitor->SetMappedValue('sHtmlHiddenFields', $sHtmlHiddenFields);
-    }
-
-    /**
-     * @return string
-     */
-    private function getMediaManagerButtonHtml(): string
-    {
-        $mediaManagerUrl = $this->mediaManagerUrlGenerator->getStandaloneMediaManagerUrl();
-        $buttonTitle = $this->translator->trans('chameleon_system_core.link.open_media_manager');
-
-        if (true === $this->mediaManagerUrlGenerator->openStandaloneMediaManagerInNewWindow()) {
-            $onClickEvent = "window.open('".$mediaManagerUrl."','mediaManager','_blank'); return false;";
-
-            return TCMSRender::DrawButton($buttonTitle, '#', URL_CMS.'/images/icons/image.gif', null, null, null, $onClickEvent);
-        } else {
-            return TCMSRender::DrawButton($buttonTitle, $mediaManagerUrl, URL_CMS.'/images/icons/image.gif', null, null, null,null, '_blank');
-        }
     }
 }
