+++ src/MediaManagerBundle/Bridge/Chameleon/Mapper/TCMSMediaFieldMapperDecorator.php
@@ -1,0 +1,73 @@
+<?php
+
+/*
+ * This file is part of the Chameleon System (https://www.chameleonsystem.com).
+ *
+ * (c) ESONO AG (https://www.esono.de)
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace ChameleonSystem\MediaManagerBundle\Bridge\Chameleon\Mapper;
+
+use ChameleonSystem\CoreBundle\Interfaces\MediaManagerUrlGeneratorInterface;
+use IMapperCacheTriggerRestricted;
+use IMapperRequirementsRestricted;
+use IMapperVisitorRestricted;
+use Symfony\Component\Translation\TranslatorInterface;
+
+class TCMSMediaFieldMapperDecorator implements \IViewMapper
+{
+    /**
+     * @var \TCMSMediaFieldMapper
+     */
+    private $subject;
+
+    /**
+     * @var MediaManagerUrlGeneratorInterface
+     */
+    private $mediaManagerUrlGenerator;
+
+    /**
+     * @var TranslatorInterface
+     */
+    private $translator;
+
+    public function __construct(
+        \TCMSMediaFieldMapper $subject,
+        MediaManagerUrlGeneratorInterface $mediaManagerUrlGenerator,
+        TranslatorInterface $translator)
+    {
+        $this->subject = $subject;
+        $this->mediaManagerUrlGenerator = $mediaManagerUrlGenerator;
+        $this->translator = $translator;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function Accept(IMapperVisitorRestricted $oVisitor, $bCachingEnabled, IMapperCacheTriggerRestricted $oCacheTriggerManager)
+    {
+        $this->subject->Accept($oVisitor, $bCachingEnabled, $oCacheTriggerManager);
+
+        $mediaManagerButtonHtml = $this->getMediaManagerButtonHtml();
+        $oVisitor->SetMappedValue('sHtmlManageMediaButton', $mediaManagerButtonHtml);
+    }
+
+    private function getMediaManagerButtonHtml(): string
+    {
+        $mediaManagerUrl = $this->mediaManagerUrlGenerator->getStandaloneMediaManagerUrl();
+        $buttonTitle = $this->translator->trans('chameleon_system_core.link.open_media_manager');
+
+        return \TCMSRender::DrawButton($buttonTitle, $mediaManagerUrl, URL_CMS.'/images/icons/image.gif', null, null, null, null, '_blank');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function GetRequirements(IMapperRequirementsRestricted $oRequirements)
+    {
+        $this->subject->GetRequirements($oRequirements);
+    }
+}
